+++ Common/GPU/OpenGL/GLRenderManager.cpp
@@ -41,77 +41,6 @@ GLRTexture::~GLRTexture() {
 	}
 }
 
-void GLDeleter::Take(GLDeleter &other) {
-	_assert_msg_(IsEmpty(), "Deleter already has stuff");
-	shaders = std::move(other.shaders);
-	programs = std::move(other.programs);
-	buffers = std::move(other.buffers);
-	textures = std::move(other.textures);
-	inputLayouts = std::move(other.inputLayouts);
-	framebuffers = std::move(other.framebuffers);
-	pushBuffers = std::move(other.pushBuffers);
-	other.shaders.clear();
-	other.programs.clear();
-	other.buffers.clear();
-	other.textures.clear();
-	other.inputLayouts.clear();
-	other.framebuffers.clear();
-	other.pushBuffers.clear();
-}
-
-// Runs on the GPU thread.
-void GLDeleter::Perform(GLRenderManager *renderManager, bool skipGLCalls) {
-	for (auto pushBuffer : pushBuffers) {
-		renderManager->UnregisterPushBuffer(pushBuffer);
-		if (skipGLCalls) {
-			pushBuffer->Destroy(false);
-		}
-		delete pushBuffer;
-	}
-	pushBuffers.clear();
-	for (auto shader : shaders) {
-		if (skipGLCalls)
-			shader->shader = 0;  // prevent the glDeleteShader
-		delete shader;
-	}
-	shaders.clear();
-	for (auto program : programs) {
-		if (skipGLCalls)
-			program->program = 0;  // prevent the glDeleteProgram
-		delete program;
-	}
-	programs.clear();
-	for (auto buffer : buffers) {
-		if (skipGLCalls)
-			buffer->buffer_ = 0;
-		delete buffer;
-	}
-	buffers.clear();
-	for (auto texture : textures) {
-		if (skipGLCalls)
-			texture->texture = 0;
-		delete texture;
-	}
-	textures.clear();
-	for (auto inputLayout : inputLayouts) {
-		// No GL objects in an inputLayout yet
-		delete inputLayout;
-	}
-	inputLayouts.clear();
-	for (auto framebuffer : framebuffers) {
-		if (skipGLCalls) {
-			framebuffer->handle = 0;
-			framebuffer->color_texture.texture = 0;
-			framebuffer->z_stencil_buffer = 0;
-			framebuffer->z_stencil_texture.texture = 0;
-			framebuffer->z_buffer = 0;
-			framebuffer->stencil_buffer = 0;
-		}
-		delete framebuffer;
-	}
-	framebuffers.clear();
-}
-
 GLRenderManager::~GLRenderManager() {
 	_dbg_assert_(!run_);
 
@@ -414,7 +343,7 @@ void GLRenderManager::BeginFrame() {
 
 	int curFrame = GetCurFrame();
 
-	FrameData &frameData = frameData_[curFrame];
+	GLFrameData &frameData = frameData_[curFrame];
 	{
 		VLOG("PUSH: BeginFrame (curFrame = %d, readyForFence = %d, time=%0.3f)", curFrame, (int)frameData.readyForFence, time_now_d());
 		std::unique_lock<std::mutex> lock(frameData.fenceMutex);
@@ -435,7 +364,7 @@ void GLRenderManager::Finish() {
 	curRenderStep_ = nullptr;  // EndCurRenderStep is this simple here.
 
 	int curFrame = GetCurFrame();
-	FrameData &frameData = frameData_[curFrame];
+	GLFrameData &frameData = frameData_[curFrame];
 
 	frameData_[curFrame].deleter.Take(deleter_);
 
@@ -463,7 +392,7 @@ void GLRenderManager::Finish() {
 
 // Render thread. Returns true if the caller should handle a swap.
 bool GLRenderManager::Run(GLRRenderThreadTask &task) {
-	FrameData &frameData = frameData_[task.frame];
+	GLFrameData &frameData = frameData_[task.frame];
 
 	if (!frameData.hasBegun) {
 		frameData.hasBegun = true;
